requirements.txt
fastapi
uvicorn
selenium
sqlalchemy
pydantic
beautifulsoup4
celery
redis
python-dotenv
sentry-sdk